+++ source/Shared/Tests/Fixtures/DomainTestDataFactory.cs
@@ -13,7 +13,9 @@
 // limitations under the License.
 
 using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
+using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationDescription;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
+using NodaTime;
 
 namespace Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures;
 
@@ -39,6 +41,77 @@ public static class DomainTestDataFactory
             "0CBB5427-E406-4E6F-A975-E21A69A3CA2A");
     }
 
+    public static OrchestrationDescription CreateOrchestrationDescription(
+        string? recurringCronExpression = default)
+    {
+        var orchestrationDescription = new OrchestrationDescription(
+            uniqueName: new OrchestrationDescriptionUniqueName("TestOrchestration", 4),
+            canBeScheduled: true,
+            functionName: "TestOrchestrationFunction");
+
+        if (recurringCronExpression != null)
+            orchestrationDescription.RecurringCronExpression = recurringCronExpression;
+
+        orchestrationDescription.ParameterDefinition.SetFromType<OrchestrationParameter>();
+
+        orchestrationDescription.AppendStepDescription("Test step 1");
+        orchestrationDescription.AppendStepDescription("Test step 2");
+        orchestrationDescription.AppendStepDescription("Test step 3", canBeSkipped: true, skipReason: "Because we are testing");
+
+        return orchestrationDescription;
+    }
+
+    public static OrchestrationInstance CreateOrchestrationInstance(
+        OrchestrationDescription orchestrationDescription,
+        OperatingIdentity? identity = default,
+        Instant? runAt = default,
+        int? testInt = default,
+        IdempotencyKey? idempotencyKey = default)
+    {
+        var operatingIdentity = identity ?? EnergySupplier.UserIdentity;
+
+        var orchestrationInstance = OrchestrationInstance.CreateFromDescription(
+            operatingIdentity,
+            orchestrationDescription,
+            skipStepsBySequence: [3],
+            clock: SystemClock.Instance,
+            runAt: runAt,
+            idempotencyKey: idempotencyKey,
+            actorMessageId: new ActorMessageId(Guid.NewGuid().ToString()),
+            transactionId: new TransactionId(Guid.NewGuid().ToString()),
+            meteringPointId: new MeteringPointId(Guid.NewGuid().ToString()));
+
+        orchestrationInstance.CustomState.SetFromInstance(new OrchestrationInstanceCustomState
+        {
+            TestId = Guid.NewGuid(),
+            TestString = "Something new",
+        });
+
+        orchestrationInstance.ParameterValue.SetFromInstance(new OrchestrationParameter
+        {
+            TestString = "Test string",
+            TestInt = testInt ?? 42,
+        });
+
+        return orchestrationInstance;
+    }
+
+    public class OrchestrationParameter
+    {
+        public string? TestString { get; set; }
+
+        public int? TestInt { get; set; }
+    }
+
+    public class OrchestrationInstanceCustomState
+    {
+        public Guid TestId { get; set; }
+
+        public string? TestString { get; set; }
+    }
+
+    #region IdentityTuple
+
     public sealed record IdentityTuple
     {
         public IdentityTuple(
@@ -60,4 +133,6 @@ public static class DomainTestDataFactory
 
         public ActorIdentity ActorIdentity { get; }
     }
+
+    #endregion
 }
